Enabled InterfaceScreen() to recompose if isPlayerNearby is true
Created SwitchPlayerNearbyBooleanButton() and switchPlayerNearby() to create a new button on the InterfaceScreen() that switches the value of isPlayerNearby from false to true and vice-versa. This allowed me to test if the screen will update and now it does! Should work when a player is actually nearby but this still needs to be tested and some of the code changed.